make default "get" command 'cp  --reflink=auto' to make files CoW if possible.
remove check for writeable output file if output filename is '-'.

add sudo access for get command.  all operations that might require root
priviledges will be called with 'sudo' if the calling user of snapsched isn't
root uid to begin with.  this allows normal users to do a get if they have the
necessary read and write priviledges.

add mod_maxsnapcount_help function to snapsched.

make a separate get_help function for ssched_get, explaining a little more.
especially since the args are not the same as ssched_diff as the diff help
claimed.